+++ lib/src/i18n/has_directional_safe_html.dart
@@ -1,11 +1,18 @@
 //Copyright (C) 2012 Sergey Akopkokhyants. All Rights Reserved.
 //Author: akserg
 
-part of dart_web_toolkit_event;
+part of dart_web_toolkit_i18n;
 
 /**
  * An object that implements this interface contains html that has a direction.
  */
 abstract class HasDirectionalSafeHtml implements HasDirectionalText, HasSafeHtml {
 
+//  /**
+//   * Sets this object's html, also declaring its direction.
+//   *
+//   * @param html the object's new html
+//   * @param dir the html's direction
+//   */
+//  void setHtml(SafeHtml html, Direction dir);
 }